Rename package to dw5
